Move Frame::to_pandas() into a separate file
