+++ docs/api/frame/to_pandas.rst
@@ -1,5 +1,5 @@
 
 .. xmethod:: datatable.Frame.to_pandas
-    :src: src/core/frame/to_numpy.cc Frame::to_pandas
-    :doc: src/core/frame/to_numpy.cc doc_to_pandas
+    :src: src/core/frame/to_pandas.cc Frame::to_pandas
+    :doc: src/core/frame/to_pandas.cc doc_to_pandas
     :tests: tests/frame/test-to-pandas.py
